move layer determination and point shuffling to types module
